+++ includes/app_version.php
@@ -2,7 +2,7 @@
 /*
  * ITFlow
  * This file defines the current ITFlow release/version
- * Update this file each time we merge develop into master. Format is YY.M (add a .v if there is more than one release a month.
+ * Update this file each time we merge develop into master. Format is YY.MM (add a .v if there is more than one release a month.
  */
 
-DEFINE("APP_VERSION", "25.01");
+DEFINE("APP_VERSION", "25.01.1");