+++ tickets.php
@@ -280,7 +280,7 @@ $user_active_assigned_tickets = intval($row['total_tickets_assigned']);
                                     <option value="unassigned" <?php if ($ticket_assigned_filter_id == "0") { echo "selected"; } ?>>Unassigned</option>
 
                                     <?php
-                                    $sql_assign_to = mysqli_query($mysqli, "SELECT * FROM users WHERE user_archived_at IS NULL ORDER BY user_name ASC");
+                                    $sql_assign_to = mysqli_query($mysqli, "SELECT * FROM users WHERE user_type = 1 AND user_archived_at IS NULL ORDER BY user_name ASC");
                                     while ($row = mysqli_fetch_array($sql_assign_to)) {
                                         $user_id = intval($row['user_id']);
                                         $user_name = nullable_htmlentities($row['user_name']);
@@ -316,12 +316,12 @@ $user_active_assigned_tickets = intval($row['total_tickets_assigned']);
                             </th>
                             <th>
                                 <a class="text-dark" href="?<?php echo $url_query_strings_sort; ?>&sort=ticket_subject&order=<?php echo $disp; ?>">
-                                    Subject / Tasks <?php if ($sort == 'ticket_subject') { echo $order_icon; } ?>
+                                    Subject <?php if ($sort == 'ticket_subject') { echo $order_icon; } ?>
                                 </a>
                             </th>
                             <th>
                                 <a class="text-dark" href="?<?php echo $url_query_strings_sort; ?>&sort=client_name&order=<?php echo $disp; ?>">
-                                    Client / Contact <?php if ($sort == 'client_name') { echo $order_icon; } ?>
+                                    Client / <span class="text-secondary">Contact</span> <?php if ($sort == 'client_name') { echo $order_icon; } ?>
                                 </a>
                             </th>
                             
@@ -418,7 +418,7 @@ $user_active_assigned_tickets = intval($row['total_tickets_assigned']);
                             if (empty($contact_name)) {
                                 $contact_display = "-";
                             } else {
-                                $contact_display = "$contact_name<br><small class='text-secondary'>$contact_email</small>";
+                                $contact_display = "$contact_name";
                             }
 
                             // Get who last updated the ticket - to be shown in the last Response column