+++ user_details.php
@@ -14,7 +14,7 @@ require_once "includes/inc_all_user.php";
 
             <div class="row">
                 <div class="col-md-3 text-center">
-                    <?php if(file_exists("uploads/users/$session_user_id/$session_avatar")) { ?>
+                    <?php if($session_avatar) { ?>
                     <img class="img-thumbnail" src="<?php echo "uploads/users/$session_user_id/" . nullable_htmlentities($session_avatar); ?>">
                     <a href="post.php?clear_your_user_avatar&csrf_token=<?= $_SESSION['csrf_token'] ?>" class="btn btn-outline-danger btn-block">Remove Avatar</a>
                     <hr>
@@ -24,7 +24,7 @@ require_once "includes/inc_all_user.php";
                         <input type="file" class="form-control-file" accept="image/*" name="avatar">
                     </div>
                 </div>
-                <div class="col-9">
+                <div class="col-md-9">
 
                     <div class="form-group">
                         <label>Name <strong class="text-danger">*</strong></label>
@@ -56,7 +56,7 @@ require_once "includes/inc_all_user.php";
                         </div>
                     </div>
                     
-                    <button type="submit" name="edit_your_user_details" class="btn btn-primary"><i class="fas fa-check mr-2"></i>Save</button>
+                    <button type="submit" name="edit_your_user_details" class="btn btn-primary btn-responsive"><i class="fas fa-check mr-2"></i>Save</button>
 
                 </div>
             </div>